bovi-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
firebase-admin>=6.0.0
google-cloud-storage>=2.0.0
